11-7-25/ecommerce-rest-api/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.2
Pillow==9.5.0